Improvements-
- There are a few errors in the code which I have corrected for you:

The event ProductDeleted is defined with the wrong arguments. It should be defined as event ProductDeleted(uint indexed productId);.

In the deleteProduct function, the line uint256 productId = products[_index].id; is incorrect. The Product struct does not have an id field. You can remove this line.

In the deleteProduct function, the line delete productsByUser[owner][productId]; should be replaced with productsByUser[owner]--;.

The buyProduct function uses the products.length variable to check the validity of the product index. However, products.length returns the number of products in the mapping, not the highest index value. You should change the condition to require(_index < productsLength, "Invalid product index");.